builds as a static library + fixed warnings.

+++ Rongine/src/Rongine/ImGui/ImGuiLayer.cpp
@@ -68,7 +68,7 @@ namespace Rongine {
 	{
 		auto& app=Application::get();
 		ImGuiIO& io = ImGui::GetIO();
-		io.DisplaySize = ImVec2(app.getWindow().getWidth(), app.getWindow().getHeight());
+		io.DisplaySize = ImVec2((float)app.getWindow().getWidth(), (float)app.getWindow().getHeight());
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
